requirements.txt
annotated-types==0.6.0
certifi==2023.7.22
charset-normalizer==2.0.12
garth==0.4.47
idna==3.4
oauthlib==3.2.2
pycryptodome==3.19.0
pydantic==2.6.3
pydantic_core==2.16.3
requests==2.26.0
requests-oauthlib==1.3.1
typing_extensions==4.10.0
urllib3==1.26.17
stravaweblib
